requirements.txt
numpy
pynput
pillow
